Imporved code structure and updated doc strings
